Improve performance of WebSockets when there is no timeout
The common case for WebSockets is to call receive
without a timeout, or use the WebSocketResponse as
an async iterator. The read was always wrapped with
an async timeout even if the timeout was not used
which accounted for as much as 50% of the run time
each loop. If there is no timeout, avoid wrapping
the read call in the context manager which
creates a new Timeout() object each time.